docs: reorder impls on L1 Date and Edtf
